requirements.txt
langchain
langchain-community
# langchain-chroma
langchain-openai
beautifulsoup4
python-dotenv
pypdf
langgraph
langgraph-checkpoint-sqlite
tavily-python


# Will use when issues are resolved.
# langchain-unstructured
# unstructured-client
# unstructured
# "unstructured[pdf]"
# python-magic
